gpt-rag-ingestion/requirements.txt
# Web framework and async dependencies
aiohttp==3.12.14
fastapi==0.115.12
uvicorn==0.34.2
webvtt-py==0.5.1
apscheduler==3.11.0

# Azure SDK dependencies
azure-appconfiguration==1.7.1
azure-appconfiguration-provider==2.1.0
azure-identity==1.19.0
azure-keyvault-secrets==4.9.0
azure-mgmt-storage==22.1.1
azure-mgmt-web==8.0.0
azure-search-documents==11.5.2
azure-storage-blob==12.24.1
msgraph-sdk==1.5.4
azure-ai-projects==1.0.0b10
azure-cosmos==4.5.1

# Data handling and validation dependencies
jsonschema
openpyxl==3.1.5
python-dotenv==1.0.1
requests==2.32.4
tabulate==0.9.0

# AI and NLP dependencies
langchain-text-splitters==0.3.9
openai==1.55.3
tiktoken==0.7.0

# Error handling
tenacity==8.5.0

# Image library
Pillow==11.1.0

azure-monitor-opentelemetry
azure-monitor-opentelemetry-exporter
opentelemetry-instrumentation-httpx==0.52b1
opentelemetry-instrumentation-fastapi
